[Refactor] Move TypeConverter into column module
