+++ be/src/storage/schema_change_utils.cpp
@@ -18,6 +18,7 @@
 #include "column/column_helper.h"
 #include "column/column_viewer.h"
 #include "column/datum_convert.h"
+#include "column/type_converter.h"
 #include "exprs/expr_factory.h"
 #include "runtime/exec_env.h"
 #include "runtime/mem_pool.h"
